% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Carseats}
\alias{Carseats}
\title{Carseats data}
\format{
Carseats
A simulated data set with 400 observations and 11 rows
\describe{
\item{Sales}{Unit sales (in thousands) at each location}
\item{CompPrice}{Price charged by competitor at each location}
\item{Income}{Community income level (in thousands of dollars)}
\item{Advertising}{Local advertising budget for company at each location (in thousands of dollars)}
\item{Population}{Population size in region (in thousands)}
\item{Price}{Price company charges for car seats at each site}
\item{ShelveLoc}{A factor with levels Bad, Good and Medium indicating the quality of the shelving location for the car seats at each site}
\item{Age}{Average age of the local population}
\item{Urban}{A factor with levels No and Yes to indicate whether the store is in an urban or rural location}
\item{US}{A factor with levels No and Yes to indicate whether the store is in the US or not}
}
}
\source{
ISLR data set, https://www.rdocumentation.org/packages/ISLR/versions/1.4/topics/Carseats
}
\usage{
Carseats
}
\description{
This is the Carseats data as shown in the ISLR package.
}
\keyword{datasets}
