% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassificationEnsembles.R
\name{ClassificationEnsemblesDemo}
\alias{ClassificationEnsemblesDemo}
\title{Run Fast Validation Classification Demo Suite}
\usage{
ClassificationEnsemblesDemo()
}
\value{
An invisible trained \code{classification_pipeline} object run over a standard medical diagnostic set.
}
\description{
Run Fast Validation Classification Demo Suite
}
