% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dry_Beans}
\alias{Dry_Beans}
\title{Dry Beans}
\format{
Dry_Beans
}
\source{
The full data set is on UCI: https://archive.ics.uci.edu/dataset/602/dry+bean+dataset
\describe{
 \item{Area}{The area of a bean zone and the number of pixels within its boundaries}
 \item{Perimeter}{Bean circumference is defined as the length of its border}
 \item{MajorAxisLength}{The distance between the ends of the longest line that can be drawn from a bean}
 \item{MinorAxisLength}{The longest line that can be drawn from the bean while standing perpendicular to the main axis}
 \item{AspectRatio}{Defines the relationship between MajorAxisLength and MinorAxisLength}
 \item{Eccentricity}{Eccentricity of the ellipse having the same moments as the region}
 \item{ConvexArea}{Number of pixels in the smallest convex polygon that can contain the area of a bean seed}
 \item{EquivDiameter}{Equivalent diameter: The diameter of a circle having the same area as a bean seed area}
 \item{Extent}{The ratio of the pixels in the bounding box to the bean area}
 \item{Solidity}{Also known as convexity. The ratio of the pixels in the convex shell to those found in beans.}
 \item{Roundness}{Calculated with the following formula: (4piA)/(P^2)}
 \item{Compactness}{Measures the roundness of an object}
 \item{ShapeFactor1}{Continuous value}
 \item{ShapeFactor2}{Continuous value}
 \item{ShapeFactor3}{Continuous value}
 \item{ShapeFactor4}{Continuous value}
 \item{Class}{(Seker, Barbunya, Bombay, Cali, Dermosan, Horoz and Sira)}
}
}
\usage{
Dry_Beans
}
\description{
A data set of dry beans
}
\keyword{datasets}
