% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassificationEnsembles.R
\name{load_pipeline}
\alias{load_pipeline}
\title{Load and Decompress Package Pipeline Footprints}
\usage{
load_pipeline(file_path)
}
\arguments{
\item{file_path}{Character path mapping to a serialized rds file.}
}
\value{
A deserialized \code{classification_pipeline} object.
}
\description{
Load and Decompress Package Pipeline Footprints
}
