% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Maternal_Health_Risk}
\alias{Maternal_Health_Risk}
\title{Maternal Health Risk}
\format{
Maternal_Health_Risk
Age, Systolic Blood Pressure as SystolicBP, Diastolic BP as DiastolicBP, Blood Sugar as BS, Body Temperature as BodyTemp, HeartRate and RiskLevel. All these are the responsible and significant risk factors for maternal mortality, that is one of the main concern of SDG of UN.

\describe{
\item{Age}{Any ages in years when a women during pregnant.}
\item{SystolicBP}{Upper value of Blood Pressure in mmHg, another significant attribute during pregnancy.}
\item{DiastolicBP}{Lower value of Blood Pressure in mmHg, another significant attribute during pregnancy.}
\item{BS}{Blood glucose levels is in terms of a molar concentration}
\item{BodyTemp}{Body temperature in Farenheit}
\item{HeartRate}{A normal resting heart rate}
\item{RiskLevel}{Predicted Risk Intensity Level during pregnancy considering the previous attribute.}
}
}
\usage{
Maternal_Health_Risk
}
\description{
Data has been collected from different hospitals, community clinics, maternal health cares from the rural areas of Bangladesh through the IoT based risk monitoring system.
}
\keyword{datasets}
