% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassificationEnsembles.R
\name{predict.classification_pipeline}
\alias{predict.classification_pipeline}
\title{Predict with Classification Pipeline Object}
\usage{
\method{predict}{classification_pipeline}(object, newdata, model_name = "best", type = c("class", "prob"), ...)
}
\arguments{
\item{object}{A trained classification_pipeline asset.}

\item{newdata}{A data frame containing incoming data rows to score.}

\item{model_name}{Character string specifying the target model from the leaderboard to score. Default is "best".}

\item{type}{Character choice layout. Type of prediction to return: "class" for factor labels, "prob" for a probability assignment matrix. Default is "class".}

\item{...}{Additional arguments passed to underlying predict methods.}
}
\value{
A factor vector or numeric probability matrix of classifications.
}
\description{
Predict with Classification Pipeline Object
}
