% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassificationEnsembles.R
\name{print.classification_pipeline}
\alias{print.classification_pipeline}
\title{Print Classification Pipeline Summary Report}
\usage{
\method{print}{classification_pipeline}(x, ...)
}
\arguments{
\item{x}{A classification_pipeline object generated by the Classification() engine.}

\item{...}{Additional arguments passed to underlying print methods.}
}
\description{
Print Classification Pipeline Summary Report
}
