% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClassificationEnsembles.R
\name{save_pipeline}
\alias{save_pipeline}
\title{Compress and Save Trained Classification Pipeline Assets}
\usage{
save_pipeline(object, file_path)
}
\arguments{
\item{object}{A trained \code{classification_pipeline} asset.}

\item{file_path}{Character path mapping where to save the binary output.}
}
\value{
Invisible NULL.
}
\description{
Compress and Save Trained Classification Pipeline Assets
}
